+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/SemanticTokenLegendCapability.cs
@@ -7,6 +7,5 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic
 {
     public class SemanticTokenLegendCapability : DynamicCapability
     {
-
     }
 }
