+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/package.json
@@ -1,7 +1,7 @@
 {
   "name": "microsoft.aspnetcore.razor.vscode",
   "private": true,
-  "version": "5.0.0-dev",
+  "version": "0.0.1",
   "defaults": {
     "razor": "0.0.1"
   },
